+++ packages/kysely/src/index.ts
@@ -8,7 +8,8 @@ import { createPasswordAdapter } from "./passwords";
 import { createCodesAdapter } from "./codes";
 import { createApplicationsAdapter } from "./applications";
 import { createConnectionsAdapter } from "./connections";
-import { createClientsAdapter } from "./clients";
+// import { createClientsAdapter } from "./clients";
+import { createLegacyClientsAdapter } from "./legacy-clients";
 import { createKeysAdapter } from "./keys";
 import { createCustomDomainsAdapter } from "./customDomains";
 import { createBrandingAdapter } from "./branding";
@@ -38,7 +39,24 @@ export default function createAdapters(db: Kysely<Database>): DataAdapters & {
     applications: createApplicationsAdapter(db),
     branding: createBrandingAdapter(db),
     cleanup: createCleanup(db),
-    legacyClients: createClientsAdapter(db),
+    clients: {
+      create: async () => {
+        throw new Error("Not implemented");
+      },
+      get: async () => {
+        throw new Error("Not implemented");
+      },
+      list: async () => {
+        throw new Error("Not implemented");
+      },
+      remove: async () => {
+        throw new Error("Not implemented");
+      },
+      update: async () => {
+        throw new Error("Not implemented");
+      },
+    },
+    legacyClients: createLegacyClientsAdapter(db),
     codes: createCodesAdapter(db),
     connections: createConnectionsAdapter(db),
     emailProviders: createEmailProvidersAdapter(db),